create Frame + level select layout
