+++ src/pages/LevelSelectPage/LevelSelectPage.tsx
@@ -1,6 +1,8 @@
 import { green, purple } from "@mui/material/colors";
-import LevelCard from "../../components/LevelCard";
+import LevelCard from "./LevelCard";
 import Tag from "../../components/Tag";
+import { Grid, Stack } from "@mui/material";
+import Frame from "../../components/Frame";
 
 const tags = [
   {
@@ -15,11 +17,32 @@ const tags = [
 
 export default function LevelSelectPage() {
   return (
-    <LevelCard 
-      level="1. Commit and Push"
-      difficulty="Easy"
-      tags={tags.map((tag, index) => {
-        return <Tag key={index} color={tag.color}>{tag.name}</Tag>
-      })} />
+    <Grid container spacing={2} sx={{
+      
+    }}>
+      <Grid item xs={4}>
+        <Stack direction="column">
+          <LevelCard 
+            level="1. Commit and Push"
+            difficulty="Easy"
+            tags={tags.map((tag, index) => {
+              return <Tag key={index} color={tag.color}>{tag.name}</Tag>
+            })} />
+          <LevelCard 
+            level="1. Commit and Push"
+            difficulty="Easy"
+            tags={tags.map((tag, index) => {
+              return <Tag key={index} color={tag.color}>{tag.name}</Tag>
+            })} />
+        </Stack>
+      </Grid>
+      <Grid item xs={8}>
+        <Frame sx={{
+          // Custom styles here
+        }}>
+
+        </Frame>
+      </Grid>
+    </Grid>
   );
 }
